backend: Refactoring directory module WIP
