+++ src-tauri/src/directory/crud.rs
@@ -1,7 +1,7 @@
 use sqlx::{query_as, SqliteConnection};
 
-use super::models::Directory;
-use crate::workspace::crud::CRUD;
+use super::{Directory, DirectoryCreateUpdate};
+use crate::workspace::crud::{CRUDHelperCreate, CRUD};
 
 impl CRUD for Directory {
     type Model = Directory;
@@ -21,5 +21,15 @@ impl CRUD for Directory {
         .bind(pk)
         .fetch_one(db_connection)
         .await
+    }
+}
+
+impl CRUDHelperCreate for DirectoryCreateUpdate {
+    fn table_name() -> String {
+        "directory".to_string()
+    }
+
+    fn get_column_names_values(&self) -> Vec<(String, crate::workspace::crud::InputValue)> {
+        todo!()
     }
 }
